Use navigation for the camera
