--- src/components/camera/PhotoActionButtons.tsx
@@ -1,87 +1,0 @@
-import React from "react";
-import {
-  StyleSheet,
-  TouchableOpacity,
-  View,
-  ImageBackground,
-} from "react-native";
-import { uploadPhoto } from "../../firebase/storage";
-import { addPhotoToEvent } from "../../firebase/db";
-import Icon from "react-native-vector-icons/FontAwesome";
-
-export default function TakePhotoButton({
-  camera,
-  photoPreview,
-  setPhotoPreview,
-  setPhoto,
-  photo,
-}) {
-  const capturePhoto = async () => {
-    try {
-      console.log("Taking photo...");
-      const photo = await camera.current.takePhoto();
-      setPhoto(photo);
-      setPhotoPreview(true);
-    } catch (e) {
-      console.error("Failed to take photo!", e);
-    }
-  };
-
-  return photoPreview ? (
-    <View>
-      <ImageBackground
-        style={{ height: "100%", width: "100%" }}
-        source={{ uri: photo.path }}
-      >
-        <TouchableOpacity style={styles.retake}>
-          <Icon
-            name={"remove"}
-            size={50}
-            color="red"
-            onPress={async () => setPhotoPreview(false)}
-          />
-        </TouchableOpacity>
-        <TouchableOpacity style={styles.save}>
-          <Icon
-            name={"check"}
-            size={50}
-            color="green"
-            onPress={async () => {
-              console.log(photo, "PHOTO");
-
-              const imageName = await uploadPhoto(photo.path);
-              addPhotoToEvent(imageName);
-              setPhotoPreview(false);
-            }}
-          />
-        </TouchableOpacity>
-      </ImageBackground>
-    </View>
-  ) : (
-    <TouchableOpacity style={styles.content}>
-      <Icon name={"square-o"} size={80} color="red" onPress={capturePhoto} />
-    </TouchableOpacity>
-  );
-}
-
-const styles = StyleSheet.create({
-  content: {
-    flex: 1,
-    flexDirection: "column",
-    alignItems: "center",
-    justifyContent: "flex-end",
-    padding: 50,
-  },
-  retake: {
-    display: "flex",
-    flexDirection: "row",
-    justifyContent: "center",
-    alignSelf: "flex-start",
-  },
-  save: {
-    display: "flex",
-    flexDirection: "row",
-    justifyContent: "center",
-    alignSelf: "flex-end",
-  },
-});
